pipeline-opensource/requirements.txt
autogluon.tabular[all]
dask[dataframe]
datasets
dvc[all]
dvclive
hydra_colorlog
hydra-core
joblib
matplotlib
numpy
omegaconf
pandas
scikit-learn
seaborn
shap
tensorflow
tensorflow-addons
tensorflow-cpu